fix(streams): apply the pre-gate-u review-round repairs, e19
Repairs from the three-reviewer consolidation (epoch-board
e19-review-round-consolidation.md), all gates green:

- migration 0005 pins feed cursors non-negative; the decoder comment
  now names the real enforcer (C2)
- ack advances with a plain UPDATE and a row-count assertion; the
  unreachable upsert-insert branch is gone (C5)
- needless allocations removed on both feed paths; the cursor-to-i64
  conversion asserts instead of falling back (C6)
- one DEFAULT_LOCK_TIMEOUT and one lock_timeout_ms helper serve both
  write paths (C7)
- the undelivered-ack message says exceeds, not precedes (C8)
- doc repairs: the ADR names AppendError::LockTimeout and the
  waiter's bound precisely, TransactError::LockTimeout's display
  names the writer lock, the pg feed's row-lock scope is the
  transaction's, the design record's watermark-durability wording
  matches what shipped, the spike header marks its shapes pre-pivot
  and its re-run an out-of-funnel writer, the postgres module docs
  state the funnel's operating assumptions (C9/C14/C16)
- FeedCursor::next_position removed (unused, panicked on legal
  input, doc overclaimed); constructor guards gain unit tests;
  conformance cases shed their unused Clone/'static bounds (C10/C11/C13)
- the no-latecomers property test's docstring now says what it pins
  and points at the new discriminating proof: a raw holder inside
  WRITER_LOCK blocks a rival append, a poll in the window sees only
  committed rows, and the rival commits above the holder's burned
  value after release (C1); the randomized model tracks the appended
  log independently of the feed under test (C12)
- harness noise: one pool per feed-spec wiring, no double deadlines,
  no discarded bindings (C20/C21/C22 included)

Card: E19

+++ tests/spike_ledger.rs
@@ -1,11 +1,19 @@
 //! The E19 gate-S spike: ADR 0010's validation gate, executed before
 //! the ledger is built. Both paths run as raw SQL over dedicated
 //! connections so the only delta between them is the ledger protocol
-//! itself - the unledgered path mirrors the shipped append/transact
-//! statement shapes, the ledgered path adds the allocation
+//! itself - the unledgered path mirrors the pre-pivot append/transact
+//! statement shapes (per-stream advisory locks), the ledgered path
+//! adds the allocation
 //! transaction (advisory-locked `nextval` draws plus the ledger-row
 //! insert), the first-statement claim, explicit sequence values, and
 //! the guarded COMMITTED update.
+//!
+//! Run of record for a superseded design: the spike adjudicated the
+//! ledger against the per-stream-lock write path the funnel has since
+//! replaced, so its shapes no longer mirror the shipped code, and a
+//! re-run writes `stream_events` OUTSIDE the single-writer funnel -
+//! only run it against a scratch database (it also leaves the
+//! `spike_e19_ledger` table and its `spike-*` categories behind).
 //!
 //! Pre-registered verdict (ADR 0010, applied mechanically):
 //!
@@ -330,7 +338,7 @@ async fn batch_ledgered(client: &mut Client, category: &str, streams: &[String])
                      (category, stream_key, event_type, sequence, event_data, \
                       event_metadata, global_sequence) \
                      VALUES ($1, $2, 'Spike', $3, '\"{}\"'::jsonb, '{}'::jsonb, $4)",
-                    &[&category, &key, &(head + offset as i64), &drawn],
+                    &[&category, &key, &(head + offset), &drawn],
                 )
                 .await
                 .expect("insert");
